Refactor events and add removed stuff.
